Revamp for album release
