+++ src/components/Music/Albums/ChickenScratch/ChickenScratchAlbumCard.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import AlbumCard from '../../AlbumCard';
+
+
+export default function ChickenScratchAlbumCard(props) {
+
+    return (
+        <AlbumCard front_img_url="/images/albums/chicken-scratch/CoverFront.jpg"
+            back_img_url="/images/albums/chicken-scratch/CoverBack.jpg" {...props} />)
+}
