+++ src/components/Music/Music.jsx
@@ -1,20 +1,31 @@
 import React from 'react';
 import { Helmet } from "react-helmet";
+import ChickenScratchAlbumCard from './Albums/ChickenScratch/ChickenScratchAlbumCard';
 
 function Music() {
     return (
-        <div>
+        <div className="container">
             <Helmet>
                 <title>RR - Music</title>
             </Helmet>
 
+            <h1>Music</h1>
 
-            <h1>New Music Coming Soon!</h1>
-            <p>We are releasing an album late spring of 2022. Stay up to date with us on <a href="https://www.facebook.com/RoosterRidgeOfficial/"> Facebook </a> for updates and promotions!</p>
-            <br />
+            <div className="row">
+                {albumCards.map((album) => (<a key={album.props.name} href={album.props.creditsUrl} style={{ color: "white" }} >
+                    {album}
+                </a>)
+
+                )}
+                {albumCards.length === 0 && <p className="container-fluid" style={{ color: "white", margin: "auto" }}>We don't have any public shows on the books for now. Please see our <a href="/contact">contact</a> page if you are interested in booking Rooster Ridge.</p>}
+            </div>
+
         </div>
     );
 }
 
+const albumCards = [
+    <ChickenScratchAlbumCard name="Chicken Scratch" creditsUrl="/music/chicken-scratch" />
+]
 
 export default Music
